version 2 (debugged)

+++ prepare_dataset.py
@@ -24,30 +24,36 @@ def move_off_files(src_dir, dest_dir):
         os.makedirs(dest_dir)
     
     off_files = glob(osp.join(src_dir, '**/*.off'), recursive=True)
-    with tqdm(total=len(off_files), desc='Moving datasets') as pbar:
-        for file_path in off_files:
-            dest_file_path = osp.join(dest_dir, osp.basename(file_path))
-            shutil.move(file_path, dest_file_path)
-            pbar.update(1)
+    for file_path in off_files:
+        dest_file_path = osp.join(dest_dir, osp.basename(file_path))
+        shutil.move(file_path, dest_file_path)
 
-def download_modelnet40(data_dir):
-    """Download and extract the ModelNet40 dataset."""
-    url = 'http://modelnet.cs.princeton.edu/ModelNet40.zip'
-    zip_path = osp.join(data_dir, 'ModelNet40.zip')
-
-    print("Downloading ModelNet40 dataset...")
+def download_modelnet10(data_dir):
     def show_progress(block_num, block_size, total_size):
         downloaded = block_num * block_size
         progress = downloaded / total_size * 100
         print(f"\rProgress: {progress:.2f}%", end="")
 
-    try:
-        urllib.request.urlretrieve(url, zip_path, show_progress)
-        print("\nDownload complete.")
-    except Exception as e:
-        raise Exception(f"Failed to download ModelNet40 dataset: {e}")
+    os.makedirs(data_dir, exist_ok=True)
+    """Download and extract the ModelNet10 dataset."""
+    url = 'http://3dvision.princeton.edu/projects/2014/3DShapeNets/ModelNet10.zip'
+    zip_path = osp.join(data_dir, 'ModelNet10.zip')
 
-    print("Extracting ModelNet40 dataset...")
+    if osp.exists(osp.join(data_dir, 'meshes')):
+        print("ModelNet10 dataset already exists.")
+        return
+    
+    if osp.exists(zip_path):
+        print("ModelNet10 zip file already exists.")
+    else:
+        print("Downloading ModelNet10 dataset...")
+        try:
+            urllib.request.urlretrieve(url, zip_path, show_progress)
+            print("\nDownload complete.")
+        except Exception as e:
+            raise Exception(f"Failed to download ModelNet10 dataset: {e}")
+
+    print("Extracting ModelNet10 dataset...")
     try:
         with zipfile.ZipFile(zip_path, 'r') as zip_ref:
             zip_ref.extractall(data_dir)
@@ -57,12 +63,12 @@ def download_modelnet40(data_dir):
         if not osp.exists(mesh_path):
             os.makedirs(mesh_path)
 
-        move_off_files(osp.join(data_dir, 'ModelNet40'), mesh_path)
+        move_off_files(osp.join(data_dir, 'ModelNet10'), mesh_path)
         print("Extraction complete.")
     except Exception as e:
-        raise Exception(f"Failed to extract ModelNet40 dataset: {e}")
+        raise Exception(f"Failed to extract ModelNet10 dataset: {e}")
     finally:
-        shutil.rmtree(osp.join(data_dir, 'ModelNet40'))
+        shutil.rmtree(osp.join(data_dir, 'ModelNet10'))
 
 def to_pcd(off_path, num_points):
     """Convert an OFF file to a point cloud with a given number of points."""
@@ -73,17 +79,29 @@ def load_model(config):
     """Load the model from the Hugging Face Hub."""
     model = models.make_model(config)
     model_name = "OpenShape/openshape-pointbert-vitg14-rgb"
-
+    
     checkpoint = torch.load(hf_hub_download(repo_id=model_name, filename="model.pt"))
+    pattern = re.compile(r'^module\.')
+    key_mapping = {
+        'pp_transformer.lift.0.weight': 'pp_transformer.conv.weight',
+        'pp_transformer.lift.0.bias': 'pp_transformer.conv.bias',
+        'pp_transformer.lift.2.weight': 'pp_transformer.norm.weight',
+        'pp_transformer.lift.2.bias': 'pp_transformer.norm.bias',
+        'proj.weight': 'head.weight',
+        'proj.bias': 'head.bias'
+    }
+    
     model_dict = OrderedDict()
-    pattern = re.compile('module.')
     for k, v in checkpoint['state_dict'].items():
-        if re.search("module", k):
-            model_dict[re.sub(pattern, '', k)] = v
+        new_key = re.sub(pattern, '', k).replace('ppat', 'pp_transformer')
+        if new_key in key_mapping:
+            new_key = key_mapping[new_key]
+        model_dict[new_key] = v
+
     model.load_state_dict(model_dict)
     return model
 
-def preprocess_modelnet40(model, configs, y_up=True):
+def preprocess_modelnet10(model, configs, y_up=True):
     """Convert all OFF files in the dataset to point cloud files and extract features."""
     off_files = glob(osp.join(configs.data_dir, 'meshes', '*.off'), recursive=True)
     off_files.sort()
@@ -92,8 +110,6 @@ def preprocess_modelnet40(model, configs, y_up=True):
     embedded_features = {}
     os.makedirs('./modelnet_embed', exist_ok=True)
     for off_file in tqdm(off_files, desc='Converting the dataset type'):
-        if configs.preprocessing_ckpt is not None and off_file < configs.preprocessing_ckpt:
-            continue
         pcd = to_pcd(off_file, configs.num_points)
         xyz = np.asarray(pcd.points)
         rgb = np.asarray(pcd.colors)
@@ -101,15 +117,15 @@ def preprocess_modelnet40(model, configs, y_up=True):
             xyz[:, [1, 2]] = xyz[:, [2, 1]]
         xyz = normalize_pc(xyz)
 
-        if rgb is None:
+        if rgb is None or len(rgb) == 0:
             rgb = np.ones_like(xyz) * 0.4
 
         features = np.concatenate([xyz, rgb], axis=1)
-        xyz_tensor = torch.tensor(xyz, dtype=torch.float32)
-        features_tensor = torch.tensor(features, dtype=torch.float32)
+        xyz_tensor = torch.tensor(xyz, dtype=torch.float32).unsqueeze(0).cuda()
+        features_tensor = torch.tensor(features, dtype=torch.float32).unsqueeze(0).cuda()
 
-        shape_feature = model(xyz_tensor, features_tensor, device='cuda', quantization_size=config.model.voxel_size)
-        embedded_features[off_file] = shape_feature.cpu().numpy()[0]
+        shape_feature = model(xyz_tensor, features_tensor, quantization_size=config.model.voxel_size)
+        embedded_features[off_file] = shape_feature.cpu().detach().numpy()[0]
 
         with open('./modelnet_embed/modelnet.pkl', 'wb') as f:
             pickle.dump(embedded_features, f)
@@ -117,12 +133,14 @@ def preprocess_modelnet40(model, configs, y_up=True):
     print("Preprocessing complete.")
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description='Download and preprocess ModelNet40 dataset')
-    parser.add_argument('--download_dir', type=str, default='./data', help='Directory to download ModelNet40 dataset')
+    parser = argparse.ArgumentParser(description='Download and preprocess ModelNet10 dataset')
+    parser.add_argument('--download_dir', type=str, default='./data', help='Directory to download ModelNet10 dataset')
     args = parser.parse_args()
 
     config = load_config()
 
     model = load_model(config)
-    download_modelnet40(args.download_dir)
-    preprocess_modelnet40(model, config)
+    model.cuda().eval()
+    download_modelnet10(args.download_dir)
+    preprocess_modelnet10(model, config)
+    model.cpu()
